+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- **BREAKING:** Remove global `spaces` option.
 - Added `EOL` override option to `writeFile` when using `spaces`. [#89]
 
 3.0.1 / 2017-07-05